requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
requests
beautifulsoup4
lxml
pypdf
langchain-text-splitters
openai
numpy
supabase
pyyaml
